Updated macronizer class
